Refactor bosituasjon step
